Automatically pulls and reboots when car is off for new git commits

+++ launch_chffrplus.sh
@@ -6,10 +6,13 @@ fi
 
 function launch {
   # apply update
-  if [ "$(git rev-parse HEAD)" != "$(git rev-parse @{u})" ]; then
-     git reset --hard @{u} &&
-     git clean -xdf &&
-     exec "${BASH_SOURCE[0]}"
+  file="/data/no_ota_updates"
+  if ! [ -f "$file" ]; then
+    if [ "$(git rev-parse HEAD)" != "$(git rev-parse @{u})" ]; then
+       git reset --hard @{u} &&
+       git clean -xdf &&
+       exec "${BASH_SOURCE[0]}"
+    fi
   fi
 
   # no cpu rationing for now
